pos_backend_by_rk/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
httplib2==0.22.0
pydantic==2.9.2
requests==2.32.3
tenacity==9.0.0
qrcode==7.4.2
Pillow==10.4.0
